+++ crates/full-node/sov-ethereum/src/handlers/get_logs/service.rs
@@ -72,7 +72,9 @@ pub struct LogsService<S: Spec, Seq: Sequencer<Spec = S>> {
     max_logs: Immutable<usize>,
     evm: Evm<S>,
     logs: Vec<Log>,
+    logs_serialized_size: usize,
     state: ApiStateAccessor<S>,
+    response_size_limit: Immutable<usize>,
     _phantom: PhantomData<(S, Seq)>,
 }
 
@@ -88,6 +90,7 @@ where
         cursor: Option<Cursor>,
         max_logs: usize,
         state: ApiStateAccessor<S>,
+        response_size_limit: usize,
     ) -> Self {
         Self {
             filter: filter.into(),
@@ -96,6 +99,8 @@ where
             state,
             evm: Evm::<S>::default(),
             logs: vec![],
+            logs_serialized_size: 0,
+            response_size_limit: response_size_limit.into(),
             _phantom: PhantomData,
         }
     }
@@ -269,6 +274,15 @@ where
                 log_index: Some(receipt.log_index_start + idx as u64),
                 removed: false,
             };
+            let log_size = serialized_size(&rpc_log);
+            if self.logs_serialized_size + log_size >= *self.response_size_limit {
+                return Ok(Some(Cursor {
+                    block_height: header.number(),
+                    tx_index_absolute,
+                    log_index_in_tx: idx as u32,
+                }));
+            }
+            self.logs_serialized_size += log_size;
             self.logs.push(rpc_log);
         }
         Ok(None)
@@ -307,4 +321,19 @@ where
         let block_number = block_nr_or_tag.unwrap_or_default();
         Ok(self.evm.resolve_block_number(block_number, &mut self.state))
     }
+}
+
+fn serialized_size(log: &Log) -> usize {
+    b"\"{address\":".len() + 44 // 20 byte address, hex-encoded + open/close quotes and 0x
+        + b",\"data\":".len() + 4 + log.inner.data.data.len() * 2 // 32 byte data, hex-encoded + open/close quotes and 0x + quotation marks
+        + b",\"topics\":[]".len() + log.inner.topics().len() * 68 // 32 byte topic, hex-encoded + open/close quotes and 0x
+        + b",\"blockHash\":".len() + 68 // Block hash (0x-prefixed + 32 data bytes) + quotation marks
+        + b",\"transactionHash\":".len() + 68 // Transaction hash (0x-prefixed + 32 data bytes) + quotation mark
+        + b",\"blockNumber\":".len() + 14 // Assume a billion blocks (plus open/close quotes and 0x prefix)
+        + b",\"blockTimestamp\":".len() + 16 // Conservatively assume a long time (current unix timestamp is only 8 hex digits, this assumes 12)
+        + b",\"transactionIndex\":".len() + 12 // Conservatively assume millions of txs per block (plus open/close quotes and 0x prefix)    
+        + b",\"logIndex\":".len() + 14 // Conservatively assume 256 logs per tx
+        + b",\"removed\":false}".len() // false is longer than true
+                                       // See example serialized log below:
+                                       // r#"{"address":"0x0000000000000000000000000000000000000069","topics":["0x0000000000000000000000000000000000000000000000000000000000000069"],"data":"0x69","blockHash":"0x0000000000000000000000000000000000000000000000000000000000000069","blockNumber":"0x69","blockTimestamp":"0x69","transactionHash":"0x0000000000000000000000000000000000000000000000000000000000000069","transactionIndex":"0x69","logIndex":"0x69","removed":false}"#
 }