+++ crates/module-system/sov-test-utils/src/test_rollup.rs
@@ -388,6 +388,7 @@ impl<R: FullNodeBlueprint<Native> + Default + 'static> RollupBuilder<R> {
             stop_at_rollup_height: None,
             extension: Some(SeqConfigExtension {
                 max_log_limit: 20000,
+                response_size_limit: (1024 * 1024) - (1024 * 30), // Limit our response size to 1MB, leaving 30kb for headers, overhead, and misestimation.
             }),
             num_cache_warmup_workers: TEST_NUM_CACHE_WARMUP_WORKERS,
             separate_archival_db: true,