+++ crates/utils/sov-eth-client/src/rpc.rs
@@ -170,6 +170,16 @@ impl RpcClient {
             .await
             .unwrap()
     }
+
+    pub async fn get_logs_with_cursor_and_filter(
+        &self,
+        cursor_and_filter: &impl serde::Serialize,
+    ) -> LogsWithMaybeCursor {
+        self.ws
+            .request("eth_getLogsWithCursor", rpc_params![cursor_and_filter])
+            .await
+            .unwrap()
+    }
 }
 
 // Alloy pubsub client