+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2025-11-01
+- #2013 Ensure `eth_getLogsWithCursor` respects a 1MB response size limit. Fix its cursor deserialization behavior to match other chains.
 - #2006 Add associated Error type to the EVM module.
 
 # 2025-10-31